fix / remove WithTableFunction and add InsteadFunction
